// frontend/src/router/enum.ts
// export const ROLE_ADMIN = 'admin'
// export const ROLE_MANAGER = 'manager'
// export const ROLE_CUSTOMER = 'customer'
// export const ROLE_EMPLOYEE = 'employee'
// export const ROLE_ANYONE = [ROLE_ADMIN, ROLE_EMPLOYEE, ROLE_MANAGER, ROLE_CUSTOMER]

export enum Role {
  admin = 'admin',
  manager = 'manager',
  customer = 'customer',
  employee = 'employee'
}